fix: Pin번호와 화면 전환
