+++ Passrithm/.idea/deploymentTargetDropDown.xml
@@ -12,6 +12,6 @@
         </deviceKey>
       </Target>
     </targetSelectedWithDropDown>
-    <timeTargetWasSelectedWithDropDown value="2022-11-19T18:07:10.577932Z" />
+    <timeTargetWasSelectedWithDropDown value="2022-11-21T07:45:34.793506Z" />
   </component>
 </project>
