+++ Passrithm/app/src/main/java/com/example/passrithm/controller/pwlist/PasswordBoxRVAdapter.java
@@ -11,7 +11,7 @@ import androidx.arch.core.internal.SafeIterableMap;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.passrithm.R;
-import com.example.passrithm.controller.algoritmlist.AlgorithmBoxRVAdapter;
+
 
 import java.util.List;
