+++ Passrithm/app/src/main/java/com/example/passrithm/controller/pwlist/PasswordListFragment.java
@@ -1,113 +1,29 @@
 package com.example.passrithm.controller.pwlist;
 
-import static android.content.Intent.getIntent;
-
-import android.app.AlertDialog;
-import android.content.Context;
-import android.content.Intent;
-import android.graphics.Color;
 import android.os.Bundle;
-import android.provider.ContactsContract;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.EditText;
-import android.widget.ImageView;
-import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.RecyclerView;
-import androidx.viewbinding.ViewBinding;
 
 import com.example.passrithm.R;
-import com.example.passrithm.controller.MainActivity;
-import com.example.passrithm.controller.algoritmlist.AlgorithmGeneratorActivity;
-import com.example.passrithm.controller.algoritmlist.SelectedBox;
-import com.example.passrithm.controller.base.FindpasswordActivity;
-import com.example.passrithm.controller.base.LoginActivity;
-import com.example.passrithm.controller.base.UserAccount;
-import com.google.android.material.progressindicator.BaseProgressIndicatorSpec;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
-import com.google.firebase.database.DataSnapshot;
-import com.google.firebase.database.DatabaseError;
-
-import com.example.passrithm.R;
-import com.example.passrithm.controller.MainActivity;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
-
-import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class PasswordListFragment extends Fragment {
-    private TextView exportButton;
-    private TextView lockText;
-    private MainActivity mainActivity;
-    private ImageView lockButton;
     private List<PasswordBox> passwordBoxes = new ArrayList<>();
 
-    //private AlertDialog pinInputDialog;
-    private View view;
-    private DatabaseReference mDatabaseRef;
 
+    public View onCreateView(@NonNull LayoutInflater inflater, @NonNull ViewGroup container, @NonNull Bundle saveInstanceState){
+        View view=(View) inflater.inflate(R.layout.fragment_password_list, container, false);
 
-    //private MainActivity mainActivity;
-
-    @Override
-    public void onAttach(@NonNull Context context) {
-        super.onAttach(context);
-        mainActivity = (MainActivity) getActivity();
-    }
-
-
-    @Nullable
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        view = (View) inflater.inflate(R.layout.fragment_password_list, container, false);
-
-        exportButton = view.findViewById(R.id.export);
-        exportButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-//                mainActivity.FragmentChange("export");
-                Intent intent = new Intent(getActivity(), ExportActivity.class);
-                intent.addFlags(Intent.FLAG_ACTIVITY_NO_ANIMATION);
-                startActivity(intent);
-            }
-        });
         PasswordBoxRVAdapter passwordBoxRVAdapter = new PasswordBoxRVAdapter(requireContext(), passwordBoxes);
         RecyclerView passwordBox = view.findViewById(R.id.password_list_rc);
         passwordBox.setAdapter(passwordBoxRVAdapter);
-
-
-        lockButton = view.findViewById(R.id.password_lock);
-        lockText = view.findViewById(R.id.password_lock_text);
-
-        lockButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-
-
-                Intent intent = new Intent(getActivity(), PinSettingActivity.class);
-                intent.addFlags(Intent.FLAG_ACTIVITY_NO_ANIMATION);
-                startActivity(intent);
-
-
-                if (lockButton.getVisibility() == view.VISIBLE) {
-                    lockButton.setVisibility(view.GONE);
-                    lockText.setVisibility(view.GONE);
-                }
-            }
-        });
         return view;
     }
 }
-
-
-
-
