+++ Passrithm/app/src/main/res/layout/fragment_password_list.xml
@@ -3,70 +3,18 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/password_list_main_frm"
-    android:orientation="vertical">
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="top"
-        android:orientation="vertical">
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="horizontal">
-            <EditText
-                android:layout_width="0dp"
-                android:layout_weight="8.5"
-                android:layout_height="wrap_content"/>
-            <ImageView
-                android:layout_width="0dp"
-                android:layout_weight="1.5"
-                android:layout_height="match_parent"
-                android:src="@drawable/ic_search"
-                android:scaleType="fitCenter"
-                /></LinearLayout>
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="right"
-            android:text="export"
-            android:id="@+id/export"/>
-        <View
-            android:layout_width="match_parent"
-            android:layout_height="2dp"
-            android:layout_marginTop="25dp"
-            android:layout_centerInParent="true"
-            android:background="#c3ced9"/>
-    </LinearLayout>
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical"
-        android:gravity="center"
-
-        android:background="@color/white"
-        android:visibility="visible">
-
-        <ImageView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:scaleType="center"
-            android:src="@drawable/ic_passwordlist_lock"
-            android:id="@+id/password_lock"/>
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="자물쇠를 클릭하여 \n 핀번호를 입력해주세요"
-            android:textAlignment="center"
-            android:layout_gravity="center"
-            android:id="@+id/password_lock_text"/>
-
-    </LinearLayout>
-
+    android:orientation="vertical"
+    android:id="@+id/password_list_frm">
     <androidx.recyclerview.widget.RecyclerView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         tools:listitem="@layout/item_password_list"
         android:id="@+id/password_list_rc"
         android:visibility="visible"/>
+    <TextView
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:text="listview"
+        />
+
 </LinearLayout>
