+++ Passrithm/app/src/main/res/layout/fragment_password_lock.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    android:id="@+id/password_lock_frm">
+    <ImageView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:scaleType="center"
+        android:layout_gravity="center"
+        android:src="@drawable/ic_passwordlist_lock"
+        android:id="@+id/password_lock"/>
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="자물쇠를 클릭하여 \n 핀번호를 입력해주세요"
+        android:textAlignment="center"
+        android:layout_gravity="center"
+        android:id="@+id/password_lock_text"/>
+
+</LinearLayout>
