+++ Passrithm/app/src/main/res/layout/item_password_list.xml
@@ -11,7 +11,7 @@
         />
 
     <TextView
-        android:id="@+id/algolist_item_tv"
+        android:id="@+id/password_item_tv"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_alignStart="@id/pwlist_item_background"
